Sejal's suggestions
