+++ README.md
@@ -1,1 +1,8 @@
-# ui-extensibility-react-hello-world
+# Get Started with UI Extensions with React
+
+This project that provides a boilerplace code to get you started with UI extensions using React as frontend. To use this with CLI,
+
+1. Update your CLI to the latest version using `npm -g @hubspot/cli@next`
+2. Create a new project with `hs project create --repoPath "https://github.com/HubSpot/ui-extensibility-react-hello-world.git"`
+
+> The UI extensions with React are currently available only with private beta. To request access for the beta as well as relavant documentation, fill out this [form](https://forms.gle/WLfKffhoe1j5z2WRA).
