+++ hsproject.json
@@ -1,4 +1,4 @@
 {
-  "name": "React Hello World",
+  "name": "Get Started with React Extension",
   "srcDir": "src"
 }
