+++ src/app/app.json
@@ -1,6 +1,6 @@
 {
-  "name": "React Example App",
-  "description": "An example private app that contains a single card extension.",
+  "name": "Get started App with React",
+  "description": "This is an example of private app that shows a custom card on the Contact record tab built with React-based frontend. This card demonstrates simple handshake with HubSpot's serverless backend.",
   "scopes": ["crm.objects.contacts.read", "crm.objects.contacts.write"],
   "public": false,
   "extensions": {
